refactor(sequencer): move BlockId to core types
